Add Incident history reporting policy
This allows for a global setting for the number
of days that an incident will remain visible for
services on the main page, based on their created date.

Currently, the entire incident history is displayed, which
over time can get quite large. This will allow limitng
this to a sensible time period.

All incidents will still be available in the dashboard, as
well as via the API.