Add manual migration for removed notification method
Update was made to drop statping_emailer notification
method, but no DB migration was added to drop this from
the DB, which causes a segfault.

Added migration step to drop this entry if it exists. This
is wrapped in a general manualMigrations function that can
be used for any necessary future manual migrations.